+++ CHANGELOG.md
@@ -10,6 +10,13 @@
   supported (by both Hoist and React) using the familiar `@HoistComponent` decorator.
 * The default text input shown by `XH.prompt()` now has `selectOnFocus: true` and will confirm the
   user's entry on an <enter> keypress (same as clicking 'OK').
+* `stringExcludes` function added to form validation constraints.  This allows
+   an input value to block specific characters or strings, e.g. no slash "/" in a textInput for
+   a filename.
+* `constrainAll` function added to form validation constraints.  This takes another
+   constraint as its only argument, and applies that constraint to an array of values,
+   rather than just to one value.  This is useful for applying a constraint to inputs that produce
+   arrays, such as tag pickers.
 * `DateInput` will now accept LocalDates as `minDate` and `maxDate` props.
 * Individual `Buttons` within a `ButtonGroupInput` will accept a disabled prop while continuing to
   respect the overall `ButtonGroupInput`'s disabled prop.